API: Move PluginMainWidget enums to init module of spyder.api.widgets
That's because we're going to need them in other modules of
spyder.api.widgets too.
