+++ spyder/api/widgets/__init__.py
@@ -7,3 +7,17 @@
 """
 Widgets to extend Spyder through its API.
 """
+
+
+class PluginMainWidgetWidgets:
+    CornerWidget = 'corner_widget'
+    MainToolbar = 'main_toolbar_widget'
+    OptionsToolButton = 'options_button_widget'
+    Spinner = 'spinner_widget'
+
+
+class PluginMainWidgetActions:
+    ClosePane = 'close_pane'
+    DockPane = 'dock_pane'
+    UndockPane = 'undock_pane'
+    LockUnlockPosition = 'lock_unlock_position'
